plugins/deepfakePlugin/requirements.txt
deepface
opencv-python
numpy
Pillow
tf-keras
tensorflow[and-cuda]